sso-server/requirements.txt
# sso-server/requirements.txt
flask==3.0.2
gunicorn==21.2.0
pyjwt==2.8.0